+++ src/main/java/org/garen/mc/mybatis/mapper/MenuMapper.java
@@ -6,12 +6,12 @@ import org.apache.ibatis.annotations.Param;
 import org.garen.mc.mybatis.domain.Menu;
 import org.garen.mc.mybatis.domain.MenuExample;
 
-public interface MenuMapper <T,Q,PK extends Serializable> extends CommonMapper<T, Q,PK>{
+public interface MenuMapper<T,Q,PK extends Serializable> extends CommonMapper<T, Q,PK> {
     long countByExample(MenuExample example);
 
     int deleteByExample(MenuExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Menu record);
 
@@ -19,7 +19,7 @@ public interface MenuMapper <T,Q,PK extends Serializable> extends CommonMapper<T
 
     List<Menu> selectByExample(MenuExample example);
 
-    Menu selectByPrimaryKey(Integer id);
+    Menu selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Menu record, @Param("example") MenuExample example);
